audito-maldito: Fixed incorrect arguments.
For whatever reason, the file path to the audito-maldito executable
appeared twice in the container arguments. This caused the flag
library to stop parsing arguments and ignore everything after the
file path. This happens because the file path does not start with
a "-", thus the flag library considers it a non-flag argument and
stops argument processing there.

I decided to add "-app-events-output" flag since the arguments
explicitly define most of the other flags' values,